requirements.txt
pyspark==3.5.0
kafka-python-ng==2.2.3
psycopg2-binary==2.9.9
